+++ runtime/binding-mqtt/src/main/java/io/aklivity/zilla/runtime/binding/mqtt/internal/config/MqttTopicConfigAdapter.java
@@ -21,6 +21,8 @@ import jakarta.json.JsonObjectBuilder;
 import jakarta.json.JsonValue;
 import jakarta.json.bind.adapter.JsonbAdapter;
 
+import io.aklivity.zilla.runtime.binding.mqtt.config.MqttTopicConfig;
+import io.aklivity.zilla.runtime.binding.mqtt.config.MqttTopicConfigBuilder;
 import io.aklivity.zilla.runtime.engine.config.ValidatorConfigAdapter;
 
 public class MqttTopicConfigAdapter implements JsonbAdapter<MqttTopicConfig, JsonObject>
